Edits to make linter happy
Edited the code to make the clippy linter happy -- removed unnecessary
clones, into_iter -> iter and boxed up a particularly large error.

+++ phylo/src/substitution_models/protein_models.rs
@@ -8,7 +8,7 @@ type ProteinFrequencyArray = [f64; 20];
 type ProteinFrequencyVector = FreqVector<20>;
 
 pub fn aminoacid_index() -> [i32; 255] {
-    let mut index = [-1 as i32; 255];
+    let mut index = [-1_i32; 255];
     for (i, char) in charify(AMINOACIDS_STR).into_iter().enumerate() {
         index[char as usize] = i as i32;
         index[char.to_ascii_lowercase() as usize] = i as i32;
